+++ examples/resources/hubspot_file_folder/import.sh
@@ -1,0 +1,1 @@
+tofu import hubspot_file_folder.assets 123456789